refactor: Remove `parseAnimationFrames()`
The `getKeyframes()` method relied on a call to `parseAnimationFrames()`, but it always resulted in a nearly-identical result: a length-2 Array with `maxHeight` as each object key, and values of `0` and the `contentHeight`.

A simple rewrite allows the removal of `parseAnimationFrames()` altogether.

Although it’s instinctual to avoid calling `reverse()` to mutate arrays, in this case it is perfectly safe, since the array is only used once per animation.